Fix File and directory exceptions argument and documentation
The command line arguments file_excep and dir_excep are not working as
intended. It is fixed now after updated nargs='*', the files and
directories are now accepted as a list.
Also updated the readme.md and the documentation in file_parser.py

Affected files: file_parser.py, readme.md